Added answers to Q3
I put "- Maybe: It provides a higher level of automation as it gives users the option to pick our algorithm that uses AI/NLP to find personally identifiable information, and anonymize it." as one of the points why people might choose our products. Feel free to remove this if it is out of scope.